Desactivation du spell checker Scayt du CKEditor (problème de confidentialité du à l'envoi de certain mot au web service) Correction de la funcExp Percentage qui génèrait un warning non numeric value.

+++ data/wigiiSystem/core/api/impl/RecordEvaluator.php
@@ -875,7 +875,7 @@ class RecordEvaluator implements FuncExpEvaluator
 		//if the evaluateArg is numeric it could have 1000 separators
 		$arg1 = str_replace("'", "", $this->evaluateArg($args[1]));
 		if($arg1!=0){
-			return (str_replace("'", "", $this->evaluateArg($args[0]))/$arg1)*100;
+			return ((float)str_replace("'", "", $this->evaluateArg($args[0]))/(float)$arg1)*100;
 		}
 		else return null;
 	}
